Don't call create in factories.
I learned I shouldn't be calling create in factories. This causes the factory to run even when you don't need it.

This meant that whenever I ran UserSeeder I ended up with an extra group I didn't need. I was passing the group_id of the group I'd already created, yet the factory kept making a group.

By not saying create Laravel runs the factory when it needs the value. Tt grabs the ID off of the returned value. Then you're all set.

You can also use lambdas. It's much easier to call the factory directly then wrap it in a lambda.